Changed to use a 3D array of floats, instead of ints, so that the data can be passed in efficiently, without needing to make a copy of the data as an array of floats. Changed DOS to UNIX text format.

+++ Operators/TOF_SCD/INTEG.java
@@ -9,7 +9,7 @@ import gov.anl.ipns.Util.SpecialStrings.*;
 
 public class INTEG implements Wrappable,HiddenOperator{
 
-  public int[][][] JHIST=new int[0][0][0];
+  public float[][][] JHIST=new float[0][0][0];
 
   public int NXS=0;
